Refactor Garden Data to appropriate packages
