+++ src/main/java/at/hannibal2/skyhanni/data/garden/composter/GardenComposterUpgradesData.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.data
+package at.hannibal2.skyhanni.data.garden.composter
 
 import at.hannibal2.skyhanni.data.model.ComposterUpgrade
 import at.hannibal2.skyhanni.events.InventoryFullyOpenedEvent
