+++ src/main/java/at/hannibal2/skyhanni/data/garden/crop/GardenCropMilestonesCommunityFix.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.data
+package at.hannibal2.skyhanni.data.garden.crop
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.config.ConfigManager
